Do check whether file or folder to extract exists and only extract if it exists

+++ stages/30-chroot.sh
@@ -11,25 +11,25 @@ pre_stage() {
 # Usage: INSTALL [MODE] SOURCE DEST
 INSTALL() {
   echo -e "\033[0;32m### INSTALL $*\033[0m"
-  
+
   local src=""
   local dst=""
 
   case "$#" in
-    "2")
-      src="$1"
-      dst="$2"
-      ;;
+  "2")
+    src="$1"
+    dst="$2"
+    ;;
 
-    "3")
-      src="$2"
-      dst="$3"
-      ;;
+  "3")
+    src="$2"
+    dst="$3"
+    ;;
 
-    *)
-      echo -e "\033[0;31m### Error: INSTALL [MODE] SOURCE DEST\033[0m"
-      return 1
-      ;;
+  *)
+    echo -e "\033[0;31m### Error: INSTALL [MODE] SOURCE DEST\033[0m"
+    return 1
+    ;;
   esac
 
   if [[ -d "${src}" ]]; then
@@ -43,19 +43,21 @@ INSTALL() {
   fi
 }
 
-# EXTRACT copies a given file or directory from the image to the destination. 
+# EXTRACT copies a given file or directory from the image to the destination.
 #
 # Usage: EXTRACT SOURCE DEST
 EXTRACT() {
   echo -e "\033[0;32m### EXTRACT $*\033[0m"
-  
+
   local src="${CHROOT_MOUNT}/$1"
   local dst=$2
 
-  if [[ -d "${src}" ]]; then
-    cp -r -T -P --preserve=mode "${src}" "${dst}"
-  else
-    cp -r -P --preserve=mode "${src}" "${dst}"
+  if [[ -e "${src}" ]]; then
+    if [[ -d "${src}" ]]; then
+      cp -r -T -P --preserve=mode "${src}" "${dst}"
+    else
+      cp -r -P --preserve=mode "${src}" "${dst}"
+    fi
   fi
 }
 
@@ -89,21 +91,21 @@ ENV() {
   local value=""
 
   case "$#" in
-    "1")
-      key="$1"
-      env_vars_del "$key"
-      ;;
+  "1")
+    key="$1"
+    env_vars_del "$key"
+    ;;
 
-    "2")
-      key="$1"
-      value="$2"
-      env_vars_set "$key" "$value"
-      ;;
+  "2")
+    key="$1"
+    value="$2"
+    env_vars_set "$key" "$value"
+    ;;
 
-    *)
-      echo -e "\033[0;31m### Error: ENV KEY [VALUE]\033[0m"
-      return 1
-      ;;
+  *)
+    echo -e "\033[0;31m### Error: ENV KEY [VALUE]\033[0m"
+    return 1
+    ;;
   esac
 
   echo -e "\033[0;32m### ENV ${key}=${value}\033[0m"
@@ -123,10 +125,10 @@ RUN() {
   local cmd_env_subst="$(env_vars_subst "$cmd_esceval")"
 
   PATH=${GUEST_PATH} chroot "${CHROOT_MOUNT}" \
-    /bin/sh -c "cd ${WORKDIR_PATH}; $(env_vars_export_cmd) ${cmd_env_subst}"
+  /bin/sh -c "cd ${WORKDIR_PATH}; $(env_vars_export_cmd) ${cmd_env_subst}"
 }
 
-# HOST executed a command on the local host and can be used to prepare files, 
+# HOST executed a command on the local host and can be used to prepare files,
 # cross-compile software, etc.
 #
 # Usage: HOST CMD PARAMS...
